refactor : 해더 구조 수정
