+++ src/1_Page/Championship/ui/InfoHeader/ui/EndChampionshipModal/model/useManageSearchPlayer.ts
