+++ src/1_Page/Championship/ui/InfoHeader/ui/EndChampionshipModal/type.d.ts
@@ -1,0 +1,3 @@
+type EndChampionshipModalProps = {
+  handleToggleModal: () => void;
+};
